Extract Msf::DBManager::Import::Burp
MSP-11124
